% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{efa_pca}
\alias{efa_pca}
\title{PCA on EFA coefficients and optionally store in SFE}
\usage{
efa_pca(sfe, efa_dim_name = "EFA", n_pcs = 10, store_name = "EFA_PCA")
}
\arguments{
\item{sfe}{SFE object with EFA reducedDim already stored}

\item{efa_dim_name}{Name of the EFA reducedDim}

\item{n_pcs}{Number of PCs to retain}

\item{store_name}{Name for the PCA reducedDim (NULL = don't store)}
}
\value{
prcomp result, invisibly
}
\description{
PCA on EFA coefficients and optionally store in SFE
}
