% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{get_nb}
\alias{get_nb}
\title{Extract the nb (neighbors) list from a colGraph}
\usage{
get_nb(sfe, graph_name = "knn")
}
\arguments{
\item{sfe}{SpatialFeatureExperiment object}

\item{graph_name}{Name of the colGraph}
}
\value{
A spdep nb object
}
\description{
Extract the nb (neighbors) list from a colGraph
}
