% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{nn_dists}
\alias{nn_dists}
\title{Compute nearest-neighbor distances from each row of A to B}
\usage{
nn_dists(A, B)
}
\arguments{
\item{A, B}{(n x 2) and (m x 2) coordinate matrices}
}
\value{
numeric vector of length nrow(A)
}
\description{
Compute nearest-neighbor distances from each row of A to B
}
