% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{polygon_area}
\alias{polygon_area}
\title{Signed area of a polygon (positive if CCW)}
\usage{
polygon_area(coords)
}
\arguments{
\item{coords}{(n x 2) matrix, not closed}
}
\description{
Signed area of a polygon (positive if CCW)
}
