% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{polygon_perimeter}
\alias{polygon_perimeter}
\title{Perimeter of a polygon}
\usage{
polygon_perimeter(coords)
}
\arguments{
\item{coords}{(n x 2) matrix, not closed}
}
\description{
Perimeter of a polygon
}
