% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{reconstruct_cell}
\alias{reconstruct_cell}
\title{Reconstruct a cell boundary from its EFA coefficients}
\usage{
reconstruct_cell(x, cell_id, n_harmonics_use = NULL, n_points = 300)
}
\arguments{
\item{x}{A \code{\link{CellEFA}} object (output of compute_efa())}

\item{cell_id}{Character; which cell to reconstruct}

\item{n_harmonics_use}{How many harmonics to use in reconstruction
(NULL = all available)}

\item{n_points}{Number of points in the reconstructed contour}
}
\value{
(n_points x 2) matrix of reconstructed coordinates in the original
  tissue coordinate system
}
\description{
Reconstruct a cell boundary from its EFA coefficients
}
