% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{remap_efa_fields}
\alias{remap_efa_fields}
\title{Remap EFA field names to the convention efourier_i expects}
\usage{
remap_efa_fields(ef)
}
\arguments{
\item{ef}{A list returned by efourier or efourier_norm}
}
\value{
The same list with fields named an, bn, cn, dn, a0, c0
}
\description{
Remap EFA field names to the convention efourier_i expects
}
