% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{resample_contour}
\alias{resample_contour}
\title{Resample a contour to n equally-spaced points by arc length}
\usage{
resample_contour(coords, n = 200)
}
\arguments{
\item{coords}{(m x 2) matrix of contour coordinates}

\item{n}{target number of points}
}
\value{
(n x 2) matrix
}
\description{
Resample a contour to n equally-spaced points by arc length
}
