% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{store_efa_in_sfe}
\alias{store_efa_in_sfe}
\title{Add EFA coefficients as a reducedDim in the SFE object}
\usage{
store_efa_in_sfe(sfe, x, name = "EFA")
}
\arguments{
\item{sfe}{A SpatialFeatureExperiment object}

\item{x}{A \code{\link{CellEFA}} object (output of compute_efa())}

\item{name}{Name for the reducedDim slot (default "EFA")}
}
\value{
The modified SFE object
}
\description{
Add EFA coefficients as a reducedDim in the SFE object
}
