% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{store_gof_in_sfe}
\alias{store_gof_in_sfe}
\title{Store GoF metrics in colData of an SFE object}
\usage{
store_gof_in_sfe(sfe, gof_df, prefix = "efa_gof_")
}
\arguments{
\item{sfe}{SpatialFeatureExperiment object}

\item{gof_df}{Output of efa_goodness_of_fit()}

\item{prefix}{Column name prefix (default "efa_gof_")}
}
\value{
Modified SFE object
}
\description{
Store GoF metrics in colData of an SFE object
}
